Crate rustc_plugin

A framework for writing plugins that integrate with the Rust compiler.

Much of this library is either directly copy/pasted, or otherwise generalized from the Clippy driver: https://github.com/rust-lang/rust-clippy/tree/master/src

Structs

RustcPluginArgs

Arguments from your plugin to the rustc_plugin framework.

Enums

CrateFilter

Specification of a set of crates.

Traits

RustcPlugin

Interface between your plugin and the rustc_plugin framework.

Functions

cli_main

The top-level function that should be called in your user-facing binary.

driver_main

The top-level function that should be called by your internal driver binary.