Trait rustc_plugin::RustcPlugin

pub trait RustcPlugin: Sized {
    type Args: Serialize + DeserializeOwned;

    // Required methods
    fn version(&self) -> Cow<'static, str>;
    fn driver_name(&self) -> Cow<'static, str>;
    fn args(&self, target_dir: &Utf8Path) -> RustcPluginArgs<Self::Args>;
    fn run(
        self,
        compiler_args: Vec<String>,
        plugin_args: Self::Args
    ) -> Result<()>;
}

Interface between your plugin and the rustc-plugin framework.

Required Associated Types

type Args: Serialize + DeserializeOwned

Command-line arguments passed by the user.

Required Methods

fn version(&self) -> Cow<'static, str>

Returns the version of your plugin.

A sensible default is your plugin’s Cargo version:

env!("CARGO_PKG_VERSION").into()

fn driver_name(&self) -> Cow<'static, str>

Returns the name of your driver binary as it’s installed in the filesystem.

Should be just the filename, not the full path.

fn args(&self, target_dir: &Utf8Path) -> RustcPluginArgs<Self::Args>

Parses and returns the CLI arguments for the plugin.

fn run(self, compiler_args: Vec<String>, plugin_args: Self::Args) -> Result<()>

Executes the plugin with a set of compiler and plugin args.